F_T32247 move ReportBundle Repositories to core
- adjust use statements
